// src/common/enums/package-status-description.enum.ts
export enum PackageStatusDescriptionEnum {
  SC = 'Envío Creado',
  PU = 'Envío recolectado',
  WC = 'Envío con mensajería de entrega',
  PL = 'Procesando Envío en la ubicación',
  DE = 'Envío Entregado',
  BA = 'Dirección Incorrecta',
  CM = 'Cliente Movio la entrega',
  DM = 'Envío Dañado',
  MC = 'Codigo Erroneo',
  CA = 'Domicilio cerrado al llegar',
  MS = 'Paquete mal sorteado',
  MD = 'Ciclo del envío completado',
  NH = 'No es un hogar',
  RD = 'Envío rechazado por el destinatario',
  AR = 'Envío en el centro de distribución.',
}